app/requirements.txt
PyQt5
SQLAlchemy
lxml
pymysql
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
